Move mamba logic into subpackage
