+++ src/main/java/org/apposed/appose/Builder.java
@@ -34,6 +34,8 @@ import java.io.IOException;
 import java.net.URISyntaxException;
 
 import org.apache.commons.compress.archivers.ArchiveException;
+import org.apposed.appose.mamba.Mamba;
+import org.apposed.appose.mamba.MambaInstallException;
 
 public class Builder {
